Step 9.12: Extract IncompleteTasksModal
